Replaced int by long. Because why not? We are in the 21st century and long's are almost always better than ints.

+++ include/Discreture/Motzkin.hpp
@@ -258,6 +258,6 @@ private:
 
 }; // end class basic_motzkin_paths
 
-using motzkin_paths = basic_motzkin_paths<int>;
+using motzkin_paths = basic_motzkin_paths<long>;
 
 } // end namespace dscr;
